[refactor] Split professors endpoint
